Added Telnet.
Got basic telnet up and running. mostly. Something seems a little
odd with it right now since it is complaining that there is a
syntax error on the first line after I added a bunch of comments.
I need to fiddle a little and see what is wrong, but for now I want
to save the work as it is so I don't loose it.