Started work on AF's web server.
This is the last big piece of the Automatic Fiesta system. With
this in place, writing software that can make your nodes respond
to web requests will be a sinch! A fair bit of it can be just
pulled from TelnetFiesta, so I'm already a good ways into this
process. But because telnet and ordinary http requests function
quite differently I will have to rework a fair bit of it. But
progress none the less! It shouldn't be to much longer now before
I have a full functional system. w00t w00t!